fix: image selection arrow up/down navigation
