+++ packages/blocks/src/image-block/components/page-image-block.ts
@@ -105,6 +105,56 @@ export class ImageBlockPageComponent extends WithDisposable(ShadowlessElement) {
         addParagraph(ctx);
         return true;
       },
+      ArrowDown: ctx => {
+        const std = this._host.std;
+
+        // If the selection is not image selection, we should not handle it.
+        // eslint-disable-next-line unicorn/prefer-array-some
+        if (!std.selection.find('image')) {
+          return false;
+        }
+
+        const event = ctx.get('keyboardState');
+        event.raw.preventDefault();
+
+        std.command
+          .chain()
+          .getNextBlock({ path: this.block.blockId })
+          .inline((ctx, next) => {
+            const { nextBlock } = ctx;
+            if (!nextBlock) return;
+
+            return next({ focusBlock: nextBlock });
+          })
+          .focusBlockStart()
+          .run();
+        return true;
+      },
+      ArrowUp: ctx => {
+        const std = this._host.std;
+
+        // If the selection is not image selection, we should not handle it.
+        // eslint-disable-next-line unicorn/prefer-array-some
+        if (!std.selection.find('image')) {
+          return false;
+        }
+
+        const event = ctx.get('keyboardState');
+        event.raw.preventDefault();
+
+        std.command
+          .chain()
+          .getPrevBlock({ path: this.block.blockId })
+          .inline((ctx, next) => {
+            const { prevBlock } = ctx;
+            if (!prevBlock) return;
+
+            return next({ focusBlock: prevBlock });
+          })
+          .focusBlockEnd()
+          .run();
+        return true;
+      },
     });
   }
